Move content components out of root components folder
