+++ admin_app/src/app/content/components/PageNavigation.tsx
