+++ admin_app/src/app/content/components/SearchBar.tsx
@@ -9,10 +9,7 @@ export interface SearchBarProps {
   setSearchTerm: React.Dispatch<React.SetStateAction<string>>;
 }
 
-export const SearchBar: React.FC<SearchBarProps> = ({
-  searchTerm,
-  setSearchTerm,
-}) => {
+export const SearchBar: React.FC<SearchBarProps> = ({ searchTerm, setSearchTerm }) => {
   return (
     <Layout.FlexBox alignItems="center" sx={{ width: "100%" }}>
       <TextField
